Moved abstract model to base/model.py
