+++ src/ewatercycle/plugins/hype/model.py
@@ -3,8 +3,7 @@ import datetime
 import logging
 import shutil
 import types
-from pathlib import Path
-from typing import Any, Iterable, Optional, Tuple, Union
+from typing import Any, Iterable, Optional, Tuple
 
 import xarray as xr
 from dateutil.parser import parse
@@ -12,7 +11,7 @@ from dateutil.tz import UTC
 
 from ewatercycle import CFG
 from ewatercycle.container import VersionImages, start_container
-from ewatercycle.model import AbstractModel
+from ewatercycle.base.model import AbstractModel
 from ewatercycle.base.parameter_set import ParameterSet
 from ewatercycle.plugins.hype.forcing import HypeForcing
 from ewatercycle.util import geographical_distances, get_time, to_absolute_path
