+++ src/ewatercycle/plugins/wflow/model.py
@@ -12,7 +12,7 @@ from cftime import num2date
 
 from ewatercycle import CFG
 from ewatercycle.container import VersionImages, start_container
-from ewatercycle.model import AbstractModel
+from ewatercycle.base.model import AbstractModel
 from ewatercycle.base.parameter_set import ParameterSet
 from ewatercycle.parametersetdb.config import CaseConfigParser
 from ewatercycle.plugins.wflow.forcing import WflowForcing
